chore: 통합/유닛 분리
